Rename login.rs to handler.rs
